# acceleration/api/openapi.yaml
openapi: 3.0.3
info:
  title: Model Router
  version: 0.2.0
  description: >
    Routes speech-to-text and text-to-speech traffic across providers and reports
    what it cost. Every path is scoped by modality, so the same provider serving
    two modalities is reported on separately. The caller is identified by a
    trusted X-Customer-Id header; real authentication is not part of this version.
servers:
  - url: http://localhost:8080
security:
  - CustomerId: []
paths:
...
  /v1/phone/vendors:
    get:
      operationId: listPhoneVendors
      summary: List the telephony vendors and whether they can be used
      description: >
        Every vendor this service knows about. A vendor that is declared but not
        implemented is listed rather than hidden, so what is missing is visible
        before a number is bought.
      responses:
        "200":
          description: The declared vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PhoneVendor"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/phone/numbers/available:
    get:
      operationId: searchPhoneNumbers
      summary: Search for numbers to buy, at one vendor or all of them
      description: >
        Naming a vendor searches only that one. Leaving it out asks every vendor
        that has its credentials, at once, and merges what they offer cheapest
        first. Vendors do not agree on how a search can be narrowed, so one whose
        API cannot express a filter is reported in `skipped` rather than asked
        without it, which would answer a search for one place with numbers from
        another.
      parameters:
        - name: vendor
          in: query
          required: false
          description: One vendor to search. Absent searches every usable vendor.
          schema:
            type: string
          example: telnyx
        - name: country
          in: query
          required: true
          description: ISO 3166-1 alpha-2 country code.
          schema:
            type: string
          example: US
        - name: area_code
          in: query
          required: false
          schema:
            type: string
        - name: contains
          in: query
          required: false
          description: Digits the number must contain, anywhere in it.
          schema:
            type: string
        - name: prefix
          in: query
          required: false
          description: >
            Digits the number must start with, matched after the country dial
            code. This differs from `contains` in where the digits have to fall.
          schema:
            type: string
          example: "719"
        - name: locality
          in: query
          required: false
          description: A city, region or rate centre.
          schema:
            type: string
          example: Denver
        - name: administrative_area
          in: query
          required: false
          description: A US state or Canadian province.
          schema:
            type: string
          example: CO
        - name: number_type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PhoneNumberType"
        - name: features
          in: query
          required: false
          description: >
            Capabilities every number must have. Repeat the parameter to require
            several. A vendor that cannot filter on one still reports what its
            numbers carry, so these are checked on the results either way.
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: "#/components/schemas/PhoneCapability"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: What the vendors are offering, and which could not answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NumberSearchResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/phone/numbers:
    get:
      operationId: listPhoneNumbers
      summary: The numbers the calling customer holds
      parameters:
        - name: include_released
          in: query
          required: false
          description: >
            Include numbers that have been given back. A released number keeps its
            row, because what it cost while it was held is still part of that
            month's bill.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: The customer's numbers, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PhoneNumber"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: buyPhoneNumber
      summary: Buy a number, which starts its monthly charge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyNumberRequest"
      responses:
        "201":
          description: The number is bought and recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumber"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/phone/numbers/{e164}:
    delete:
      operationId: releasePhoneNumber
      summary: Give a number back, which stops its monthly charge
      parameters:
        - name: e164
          in: path
          required: true
          description: The number in +15551234567 form.
          schema:
            type: string
          example: "+15125551234"
      responses:
        "204":
          description: The number was released
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/phone/numbers/{e164}/attach:
    post:
      operationId: attachPhoneNumber
      summary: Point a number at a Stream call
      description: >
        Creates the SIP inbound trunk and routing rule and tells the vendor to send
        calls there. This is what turns a bought number into one that reaches an
        agent.
      parameters:
        - name: e164
          in: path
          required: true
          schema:
            type: string
          example: "+15125551234"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachNumberRequest"
      responses:
        "200":
          description: The number now reaches a call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachedNumber"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/phone/calls:
    post:
      operationId: placePhoneCall
      summary: Place an outbound call and bridge it into a Stream call
      description: >
        Stream's SIP is inbound only, so the vendor originates the call and connects
        it to a trunk the agent is already on, rather than Stream dialling out.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceCallRequest"
      responses:
        "202":
          description: The vendor is placing the call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlacedCall"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/phone/calls/transfer:
    post:
      operationId: transferPhoneCall
      summary: Bring a human onto a call that is already happening
      description: >
        Stream's SIP is inbound only, so a transfer is a second leg rather than a
        handover: the vendor dials the human and the answered leg is routed into the
        same Stream call, after which the agent can leave. The caller is never moved,
        so nothing is lost if nobody answers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferCallRequest"
      responses:
        "202":
          description: The vendor is dialling the human
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlacedCall"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/phone/calls/{vendor_call_id}/digits:
    post:
      operationId: pressPhoneDigits
      summary: Press digits on a call placed from here
      description: >
        For getting past a menu on an outbound call. The call is named by the id its
        vendor gave when it was dialled, so only calls placed from this service can be
        pressed at. Not every vendor can do this without ending the call it is on, and
        one that cannot says so.
      parameters:
        - name: vendor_call_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PressDigitsRequest"
      responses:
        "204":
          description: The digits were pressed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    CustomerId:
      type: apiKey
      in: header
      name: X-Customer-Id
      description: Trusted customer identifier. Every statistic is keyed by it.
...
  responses:
    BadRequest:
      description: The request was malformed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: The customer header is missing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such modality, provider or shortcut
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PhoneVendor:
      type: object
      required: [vendor, implemented, ready, capabilities]
      properties:
        vendor:
          type: string
          example: twilio
        implemented:
          type: boolean
          description: Whether this service can actually work with the vendor.
        ready:
          type: boolean
          description: Implemented and holding every credential it needs.
        operations:
          type: array
          description: >
            What this service can do at the vendor. Eight vendors buy numbers and
            two of those also bridge calls, so a number is not bought from a
            vendor that cannot answer on it by accident.
          items:
            $ref: "#/components/schemas/PhoneOperation"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/PhoneCapability"
        missing_credentials:
          type: array
          description: The environment variables the vendor needs and does not have.
          items:
            type: string
...
    PhoneCapability:
      type: string
      description: >
        What a number can carry. The names are Telnyx's feature names, because
        they are the widest vocabulary any of these vendors offers.
      enum:
        [
          voice,
          sms,
          mms,
          fax,
          emergency,
          hd_voice,
          international_sms,
          local_calling,
        ]
    PhoneNumberType:
      type: string
      description: What kind of number it is, which decides who pays for the call.
      enum: [local, toll_free, mobile]
    NumberSearchResult:
      type: object
      required: [numbers, skipped]
      properties:
        numbers:
          type: array
          description: What the vendors are offering, cheapest first.
          items:
            $ref: "#/components/schemas/AvailableNumber"
        skipped:
          type: array
          description: >
            Vendors that were not part of the answer. A search that reached two
            of eight vendors found what two vendors had, and deciding whether to
            buy needs to know which.
          items:
            $ref: "#/components/schemas/SkippedVendor"
    SkippedVendor:
      type: object
      required: [vendor, reason]
      properties:
        vendor:
          type: string
          example: twilio
        reason:
          type: string
          example: cannot search by administrative_area
...
    PhoneNumber:
      type: object
      required: [e164, vendor, country, capabilities, monthly_cost_micros, purchased_at]
      properties:
        e164:
          type: string
        vendor:
          type: string
        country:
          type: string
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/PhoneCapability"
        monthly_cost_micros:
          type: integer
          format: int64
        tags:
          type: object
          additionalProperties:
            type: string
          description: The customer's own cost labels.
        stream_trunk_id:
          type: string
          description: The SIP trunk calls to this number arrive on. Absent until attached.
        purchased_at:
          type: string
          format: date-time
        released_at:
          type: string
          format: date-time
          nullable: true
    BuyNumberRequest:
      type: object
      required: [vendor, e164]
      properties:
        vendor:
          type: string
          example: twilio
        e164:
          type: string
          example: "+15125551234"
        country:
          type: string
          description: >
            The country the number was offered from, as the search reported it.
            Most vendors buy by number alone; the few that buy out of a country's
            inventory need this, and it cannot be guessed back out of the number.
          example: US
        tags:
          type: object
          additionalProperties:
            type: string
          description: Cost labels carried onto the purchase's request row.
    AttachNumberRequest:
      type: object
      properties:
        call_id:
          type: string
          description: >
            The call every caller joins. Omit to give each caller their own call,
            named after the number they rang.
        call_type:
          type: string
          description: The Stream call type. Omit for "default".
        allowed_ips:
          type: array
          description: The vendor's signalling addresses, as IPs or CIDR blocks.
          items:
            type: string
    AttachedNumber:
      type: object
      required: [trunk_id, route_id, sip_uri]
      properties:
        trunk_id:
          type: string
        route_id:
          type: string
        sip_uri:
          type: string
          description: Where the vendor sends calls, e.g. sip:trunk@sip.stream-io-api.com.
    PlaceCallRequest:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
          description: One of the customer's own numbers, which is what the person sees.
        to:
          type: string
        call_id:
          type: string
          description: >
            The Stream call the answered leg joins, and so the one the agent has to be
            in. Omit to have one named after this call, since two calls from the same
            number are two conversations.
        call_type:
          type: string
          description: The Stream call type. Omit for "default".
        ring_timeout_seconds:
          type: integer
          description: >
            How long to ring before giving up. Omit to leave the vendor's default,
            which is long enough to reach voicemail. A vendor whose call API cannot
            express it refuses the call rather than ringing for its own default.
        initial_digits:
          type: string
          description: >
            Digits pressed once the person answers, for reaching an extension behind a
            menu, e.g. "ww1234#". w is a short pause and W a long one.
        custom:
          type: object
          description: >
            Put on the Stream call, where the agent in it can read it. It is set at
            Stream rather than at the vendor, so every vendor can carry it.
          additionalProperties:
            type: string
        headers:
          type: object
          description: >
            Carried to the person's leg as custom SIP headers. Only some vendors can
            express these, and one that cannot refuses the call.
          additionalProperties:
            type: string
        tags:
          type: object
          additionalProperties:
            type: string
    PlacedCall:
      type: object
      required: [vendor_call_id, status]
      properties:
        vendor_call_id:
          type: string
        status:
          type: string
          description: The vendor's own word for where the call is, e.g. "queued".
        vendor:
          type: string
          description: Who is placing the call.
        call_id:
          type: string
          description: >
            The Stream call the answered leg is routed into. An agent that is not in it
            hears nothing when the person picks up.
        call_type:
          type: string
    TransferCallRequest:
      type: object
      required: [from, to, call_id]
      properties:
        from:
          type: string
          description: >
            The customer's number the human is dialled from, which is what they see.
        to:
          type: string
          description: The human being brought onto the call.
        call_id:
          type: string
          description: The Stream call the caller and the agent are already on.
        call_type:
          type: string
          description: The Stream call type. Omit for "default".
        tags:
          type: object
          additionalProperties:
            type: string
    PressDigitsRequest:
      type: object
      required: [vendor, digits]
      properties:
        vendor:
          type: string
          description: Who is carrying the call, e.g. "telnyx".
        digits:
          type: string
          description: >
            What to press. Only 0-9, * and # can be pressed, and w waits half a second
            between two of them.
...